Класс Table вынесен в отдельные файлы
